Exposes the version info object to users.
This will allow users to leverage the version computation logic without tightly coupling to the opinionated usage of that data.  As an example, a user could choose to have a numeric version that is an aggregate of the various parts of the version number in some way or they could order the various pieces in a unique way within the version string.