Changes custom describe processor to return an array.
Also fixes some major bugs that caused the plugin to fail to work correctly if the user didn't have a version that could be processed by the built-in versioning.  Also removes the previously baked-in support for semantic versioning and instead opts for users supplying their own versioning function if they want to go outside of the default `v1.2-tag` format.